refactor: 타입 정의 위치 변경
